fix compilation errors

+++ engine/interfaces/sc_bcp_api.cpp
@@ -503,7 +503,7 @@ bool download_item_data( item_t& item, cache::behavior_e caching )
     {
       const rapidjson::Value& sockets = js[ "socketInfo" ][ "sockets" ];
 
-      for ( rapidjson::SizeType i = 0, n = std::min( static_cast< unsigned long >( sockets.Size() ), sizeof_array( item.parsed.data.socket_color ) ); i < n; ++i )
+      for ( rapidjson::SizeType i = 0, n = std::min( static_cast< size_t >( sockets.Size() ), sizeof_array( item.parsed.data.socket_color ) ); i < n; ++i )
       {
         if ( ! sockets[ i ].HasMember( "type" ) )
           continue;
